feat: rename clients filename
